refactor: Move proposal page subcomponents into own subfolder.
