+++ src/components/Guilds/ProposalPage/ProposalInfoCard/InfoItem.tsx
