+++ src/components/Guilds/ProposalPage/ProposalInfoCard/index.tsx
