+++ src/components/Guilds/ProposalPage/ProposalVoteCard/ProposalVotes.tsx
