+++ src/components/Guilds/ProposalPage/ProposalVoteCard/index.tsx
