+++ src/components/Guilds/ProposalPage/index.tsx
